.xml files autoparsing

+++ app.js
@@ -1,10 +1,11 @@
 'use strict'
 require('dotenv').config()
 
-const fs = require("fs");
-const downloader = require("./downloader");
-const connector = require('./connector');
-const parseSchemas = require('./parseObjects');
+const fs = require("fs")
+const downloader = require("./downloader")
+const { connect, client, dbName } = require('./connector')
+const parseSchemas = require('./parseObjects')
+const { parseXml } = require('./parser')
 
 process.on('uncaughtException', (err, origin) => {
   fs.writeSync(
@@ -14,8 +15,15 @@ process.on('uncaughtException', (err, origin) => {
   );
 });
 
+require('node-schedule')
+  .scheduleJob(process.env.CRON_TIMER, function() {
+    console.log('autoparse')
+    autoParse()
+  })
+
 function runParse() {
-  if (process.env.SHOW_AVAILABLE_PARSE_SCHEMAS)  showAvailableSchemas()
+  if (process.env.CRON_PARSER) return
+  if (process.env.SHOW_AVAILABLE_PARSE_SCHEMAS) showAvailableSchemas()
   if (process.env.parseschema && process.env.filename) parseData()
   if (process.env.REMOVE_FILE) removeFile()
   if (process.env.SHOW_EXTERNAL_SOURCES) showDownloadSources()
@@ -28,14 +36,14 @@ runParse()
 
 function parseData() {
   if (!Object.keys(parseSchemas).includes(process.env.parseschema)) {
-    throw new Error ("Schema not exist. Run 'npm run showAvailableParseSchemas' to get available shemas")
+    throw new Error ("Schema does not exist. Run 'npm run showAvailableParseSchemas' to get available schemas")
   }
   fs.readdir('parse-file', (err, files) => {
     if (err) throw err
     if (!files?.includes(process.env.filename)) {
       throw new Error("File does not exit. Run 'npm run showAvailableFiles' to get available files")
     }
-    connector.connect();
+    connect();
   })
 }
 
@@ -60,14 +68,42 @@ function showDownloadSources () {
 function showFiles () {
   fs.readdir('parse-file', (err, files) => {
     console.log('Files to parse: ', files)
-  });
+  })
 }
 
-function downloadFile () {
-  if (!process.env.filename.includes('.xml')) {
+function autoParse () {
+  client.connect(async function (err) {
+    if (err) {
+      console.log(err) 
+      return
+    }
+
+    const db = client.db(dbName)
+    const files = downloader.sources.map(source => {
+      return downloadFile(source, source)
+    })
+
+    Promise.all(files).then(() => {
+      console.log('Files downloaded')
+      console.log('Parsing started')
+
+      Object.keys(parseSchemas).map(async schema => {
+        let col = db.collection(schema)
+        await col.deleteMany({})
+        parseXml(col, parseSchemas[schema].fileName + '.xml', schema)
+      })
+    })
+  })
+}
+
+function downloadFile (fileName, option) {
+  const notValidEnvName = process.env.filename && !process.env.filename.includes('.xml')
+  if (notValidEnvName && typeof fileName !== 'string') {
     throw new Error ("Specify file name 'filename=<file name with extension> npm run ...'")
   }
-  try { downloader.GET() } catch(err) {console.log(err)};
+
+  const fileNm = fileName.includes('.xml') ? fileName : fileName + '.xml'
+  return downloader.GET(fileNm, option)
 }
 
 function consoleHelp () {
